Update routing concerns
